+++ tests/src/Infolists/StateTest.php
@@ -2,9 +2,11 @@
 
 namespace Filament\Tests\Infolists;
 
+use Filament\Forms\Components\TextInput;
 use Filament\Infolists\Components\KeyValueEntry;
 use Filament\Infolists\Components\TextEntry;
 use Filament\Schemas\Components\Section;
+use Filament\Schemas\Components\Utilities\Get;
 use Filament\Schemas\Concerns\InteractsWithSchemas;
 use Filament\Schemas\Contracts\HasSchemas;
 use Filament\Schemas\Schema;
@@ -101,6 +103,93 @@ class NestedKeyValueInfolistComponent extends Component implements HasSchemas
         return <<<'BLADE'
             <div>
                 {{ $this->infolist }}
+            </div>
+            BLADE;
+    }
+}
+
+it('can render entry within form that reads form data', function (): void {
+    livewire(EntryInFormComponent::class)
+        ->assertSuccessful()
+        ->assertSchemaComponentStateSet('display', 'Current value: initial', 'form')
+        ->assertSeeText('Current value: initial')
+        ->set('data.input_field', 'updated value')
+        ->assertSchemaComponentStateSet('display', 'Current value: updated value', 'form')
+        ->assertSeeText('Current value: updated value');
+});
+
+class EntryInFormComponent extends Component implements HasSchemas
+{
+    use InteractsWithSchemas;
+
+    public array $data = [];
+
+    public function mount(): void
+    {
+        $this->form->fill([
+            'input_field' => 'initial',
+        ]);
+    }
+
+    public function form(Schema $schema): Schema
+    {
+        return $schema
+            ->statePath('data')
+            ->components([
+                TextInput::make('input_field')
+                    ->live(),
+                TextEntry::make('display')
+                    ->label('Display')
+                    ->state(fn (Get $get): string => 'Current value: ' . $get('input_field')),
+            ]);
+    }
+
+    public function render(): string
+    {
+        return <<<'BLADE'
+            <div>
+                {{ $this->form }}
+            </div>
+            BLADE;
+    }
+}
+
+it('can assert entry state within form using `assertSchemaComponentStateSet()`', function (): void {
+    livewire(EntryStateInFormComponent::class)
+        ->assertSuccessful()
+        ->assertSchemaComponentStateSet('summary', 'Form field value: test input', 'form')
+        ->assertSeeText('Form field value: test input');
+});
+
+class EntryStateInFormComponent extends Component implements HasSchemas
+{
+    use InteractsWithSchemas;
+
+    public array $data = [];
+
+    public function mount(): void
+    {
+        $this->form->fill([
+            'input' => 'test input',
+        ]);
+    }
+
+    public function form(Schema $schema): Schema
+    {
+        return $schema
+            ->statePath('data')
+            ->components([
+                TextInput::make('input'),
+                TextEntry::make('summary')
+                    ->state(fn (Get $get): string => 'Form field value: ' . $get('input')),
+            ]);
+    }
+
+    public function render(): string
+    {
+        return <<<'BLADE'
+            <div>
+                {{ $this->form }}
             </div>
             BLADE;
     }
